+++ core/src/com/github/grhscompsci2/galaga/screens/LoadingScreen.java
@@ -1,14 +1,48 @@
 package com.github.grhscompsci2.galaga.screens;
 
 import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.ScreenAdapter;
 import com.github.grhscompsci2.galaga.MyGdxGame;
 
-public class LoadingScreen extends ScreenAdapter {
+public class LoadingScreen implements Screen {
     private MyGdxGame parent;
 
     public LoadingScreen(MyGdxGame myGdxGame) {
         parent = myGdxGame;
+    }
+
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
     }
 
 }
